Numerical experiments
Now, numerical experiments can be performed, such that precipitation chemistry and/or hydrology during simulation is different from that of warm-up period.

Input file "Numexp_precipchem.txt" has the precipitation chemistry for  simulation period over which numerical experiment is to be carried out.

"Numexp_Results.txt"  file has hydrology input for simulation period over which numerical experiment is to be carried out.

It there is no "Numexp_Results.txt", hydrology in "Results.txt" file will be repeated depending on number of years for which precipitation chemistry is provided in "Numexp_precipchem.txt" file.