New structure for HBV-BioRT model
The model has been restructured so that the surface zone that was present earlier has been modified to become a simple temporary storage of water, that does not include soil moisture. The new structure is more representative of catchment processes. Please refer HBV-BioRT documentation for a better understanding of new model structure